Initial view/model split up.
